+++ pylps/solver.py
@@ -1,390 +1,267 @@
-# import copy
-# from unification import *
-
-# from pylps.constants import *
-# from pylps.exceptions import *
-# from pylps.utils import *
-
-# from pylps.config import CONFIG
-# from pylps.kb import KB
-
-# from pylps.tree_goal import TreeGoal, SolverTreeGoal
-# from pylps.lps_objects import GoalClause
-
-# from pylps.unifier import unify_fact
-# from pylps.constraints import constraints_satisfied
-
-
-# def solve_multigoal(multigoal: TreeGoal, cycle_time: int) -> bool:
-
-#     if multigoal.defer_goals:
-#         for defer_goal in multigoal.defer_goals:
-#             solve_defer_goal(defer_goal, cycle_time)
-#             if defer_goal.result is G_DISCARD:
-#                 multigoal.update_result(G_DISCARD)
-#                 return multigoal
-#             elif defer_goal.result in SOLVED_RESPONSES:
-#                 multigoal.solved_cnt += 1
-#                 multigoal.update_subs(defer_goal.new_subs)
-#             elif defer_goal.result is G_DEFER:
-#                 continue
-#             else:
-#                 # print(multigoal)
-#                 raise UnimplementedOutcomeError("multigoal_defer")
-
-#     else:
-#         cur_goal_pos = 0
-#         num_goals = len(multigoal.goals)
-#         responses = [None] * num_goals
-#         failed_prev = False
-
-#         '''
-#         DFS to move back and forth between goals
-#         This should be extracted out and hardened
-#         '''
-#         while cur_goal_pos < len(multigoal.goals):
-#             goal = multigoal.goals[cur_goal_pos]
-#             # print(goal)
-
-#             if failed_prev:
-#                 response = responses[cur_goal_pos]
-
-#                 if response.result in SOLVED_RESPONSES:
-#                     new_sub = response.get_new_sub_option()
-
-#                     if new_sub is ERROR_NO_SUB_OPTIONS:
-#                         multigoal.update_result(G_FAIL_NO_SUBS)
-#                         return
-
-#                     # Update the subs
-#                     response.update_subs(new_sub)
-#                     multigoal.update_subs(response.new_subs)
-
-#                     # Reset flags
-#                     failed_prev = False
-#                     cur_goal_pos += 1
-#                     continue
-
-#                 else:
-#                     raise UnimplementedOutcomeError("solve_multigoal_p_fail")
-
-#             # TODO: goal.goal (should be goal), in the process of switching
-#             solve_goal(goal, cycle_time)
-
-#             # print(goal)
-
-#             if goal.result is G_DISCARD:
-#                 multigoal.update_result(G_DISCARD)
-#                 return multigoal
-#             elif goal.result is G_DEFER:
-#                 multigoal.add_defer_goals(goal)
-#                 cur_goal_pos += 1
-#             elif goal.result in SOLVED_RESPONSES:
-#                 multigoal.solved_cnt += 1
-#                 multigoal.update_subs(goal.new_subs)
-#                 responses[cur_goal_pos] = copy.deepcopy(goal)
-#                 cur_goal_pos += 1
-#             elif goal.result is G_CLAUSE_FAIL:
-#                 '''
-#                 TODO: Handle clause failure correctly
-#                 Treat clause failure like a discard now
-#                 '''
-#                 goal.reset()
-#                 if cur_goal_pos == 0:
-#                     multigoal.update_result(G_CLAUSE_FAIL)
-#                     return multigoal
-
-#                 cur_goal_pos = cur_goal_pos - 1
-#                 failed_prev = True
-#             else:
-#                 # print(multigoal)
-#                 raise UnimplementedOutcomeError("solve_multigoal")
-
-#         # multigoal.update_result(G_UNSOLVED)
-
-#     if multigoal.solved_cnt == len(multigoal.goals):
-#         multigoal.update_result(G_SOLVED)
-#     elif multigoal.defer_goals:
-#         multigoal.update_result(G_DEFER)
-#     else:
-#         multigoal.update_result(G_UNSOLVED)
-
-#     return multigoal
-
-
-# def solve_goal(goal: TreeGoal, cycle_time: int) -> TreeGoal:
-#     parent = goal.parent
-#     goal.update_subs(parent.subs)
-
-#     # print(goal, subs, cycle_time)
-
-#     if goal.temporal_vars:
-#         for temporal_var in goal.temporal_vars:
-#             if temporal_var in parent.subs:
-#                 goal.temporal_sub_used = True
-
-#     KB_clauses = KB.get_clauses(goal.obj)
-
-#     if CONFIG.single_clause:
-#         KB_clauses = KB_clauses[:1]
-
-#     # Handling a complex
-#     if KB_clauses:
-#         for clause in KB_clauses:
-#             solve_goal_complex(
-#                 clause,
-#                 goal,
-#                 cycle_time
-#             )
-
-#             if goal.result is G_CLAUSE_FAIL:
-#                 goal.clear_children()
-#                 continue
-
-#             if goal.result is G_SOLVED:
-#                 KB.add_cycle_action(goal, goal.new_subs)
-#                 # print(goal)
-#                 return
-
-#             if(goal.result is G_DISCARD or
-#                     goal.result is G_DEFER):
-#                 return
-#     else:
-#         solve_goal_single(goal, cycle_time)
-#         # print(cycle_time, goal)
-#         return
-
-#     goal.clear_subs()
-#     return
-
-
-# def solve_defer_goal(goal: TreeGoal, cycle_time: int) -> TreeGoal:
-#     if goal.defer_goals:
-#         for defer_goal in goal.defer_goals:
-#             solve_defer_goal(defer_goal, cycle_time)
-
-#     else:
-#         solve_goal_single(goal, cycle_time)
-#         return
-
-#     new_defer_goals = []
-#     for defer_goal in goal.defer_goals:
-#         if defer_goal.result is G_DISCARD:
-#             goal.update_result(G_DISCARD)
-#             goal.set_defer_goals([])
-#             return goal
-#         elif defer_goal.result in SOLVED_RESPONSES:
-#             continue
-#         elif defer_goal.result is G_DEFER:
-#             new_defer_goals.append(defer_goal)
-#         else:
-#             raise UnknownOutcomeError("solve_defer_goal")
-
-#     goal.set_defer_goals(new_defer_goals)
-
-#     if not goal.defer_goals:
-#         goal.update_result(G_SOLVED)
-#     # print(cycle_time, goal)
-
-
-# def solve_goal_complex(
-#         clause: GoalClause,
-#         goal: TreeGoal,
-#         cycle_time: int) -> TreeGoal:
-
-#     clause_goal = clause.goal[0]
-#     clause_goal_object = clause_goal[0]  # This may cause issues
-#     clause_temporal_vars = tuple(
-#         var(temporal_var.name) for temporal_var in clause_goal[1:]
-#     )
-
-#     subgoal_defer = False
-
-#     # TODO: FIX / CHANGE THIS
-#     # Check if the objects match
-#     if clause_goal_object != goal.obj:
-#         return goal
-
-#     for req in clause.reqs:
-#         combined_subs = {**goal.subs, **goal.new_subs}
-#         req_goal = SolverTreeGoal(
-#             parent=goal,
-#             goal=req,
-#             children=[],
-#             subs=copy.deepcopy(combined_subs),
-#             temporal_sub_used=goal.temporal_sub_used
-#         )
-
-#         goal.add_child(req_goal)
-
-#         solve_goal_single(req_goal, cycle_time)
-
-#         if (req_goal.result is G_DISCARD or
-#                 req_goal.result is G_CLAUSE_FAIL):
-#             goal.update_result(req_goal.result)
-#             goal.clear_subs()
-#             return goal
-
-#         if req_goal.result is G_DEFER:
-#             subgoal_defer = True
-#             goal.add_defer_goals(req_goal)
-
-#         if req_goal.new_subs:
-#             goal.update_subs(req_goal.new_subs)
-
-#         goal.temporal_sub_used = req_goal.temporal_sub_used
-
-#     temporal_satisfied = goal_temporal_satisfied(goal, clause_goal)
-
-#     if not temporal_satisfied:
-#         combined_subs = {**goal.subs, **goal.new_subs}
-#         goal.temporal_satisfied = False
-
-#         children_temporal_satisifed = True
-#         for child in goal.children:
-#             if not child.temporal_satisfied:
-#                 children_temporal_satisifed = False
-
-#         if children_temporal_satisifed:
-#             reify_goal = reify(goal.temporal_vars, combined_subs)
-#             if isinstance(reify_goal[0], int):
-#                 reify_goal = (reify_goal[0], reify_goal[0] + 1)
-#                 goal.update_subs(unify(goal.temporal_vars, reify_goal))
-#                 temporal_satisfied = True
-#             else:
-#                 raise UnknownOutcomeError(reify_goal)
-#         else:
-#             raise UnknownOutcomeError('children_temporal_satisifed')
-
-#     temporal_satisfied = goal_temporal_satisfied(goal, clause_goal)
-
-#     if temporal_satisfied:
-#         combined_subs = {**goal.subs, **goal.new_subs}
-#         goal_temporal_vars = reify(clause_temporal_vars, combined_subs)
-#         goal.temporal_satisfied = True
-
-#         if strictly_increasing(goal_temporal_vars):
-#             goal.temporal_vars = goal_temporal_vars
-#             if subgoal_defer:
-#                 goal.update_result(G_DEFER)
-#             else:
-#                 goal.update_result(G_SOLVED)
-#             return
-
-#         goal.clear_subs()
-#         goal.update_result(G_DISCARD)
-#         return
-
-#     return
-
-
-# def solve_goal_single(goal: TreeGoal, cycle_time: int) -> TreeGoal:
-
-#     reify_goal = None
-#     goal.temporal_satisfied = True
-#     combined_subs = {**goal.subs, **goal.new_subs}
-
-#     if goal.temporal_vars:
-#         goal.temporal_satisfied = False
-#         if not goal.temporal_sub_used:
-#             for temporal_var in goal.temporal_vars:
-#                 if temporal_var in goal.subs:
-#                     continue
-
-#                 if not goal.temporal_sub_used:
-#                     goal.update_subs(unify(temporal_var, cycle_time))
-
-#                     # Update the combined subs
-#                     combined_subs = {**goal.subs, **goal.new_subs}
-#                     goal.temporal_sub_used = True
-
-#         if goal.obj.BaseClass is ACTION or goal.obj.BaseClass is EVENT:
-#             reify_goal = reify(goal.temporal_vars, combined_subs)
-#             if isinstance(reify_goal[0], int):
-#                 reify_goal = (reify_goal[0], reify_goal[0] + 1)
-#                 goal.update_subs(unify(goal.temporal_vars, reify_goal))
-#             else:
-#                 raise UnknownOutcomeError(reify_goal)
-#         else:
-#             raise UnknownOutcomeError("Temporal var without action")
-
-#         goal_temporal_satisfied_cnt = 0
-#         for item in reify_goal:
-#             if isinstance(item, int):
-#                 goal_temporal_satisfied_cnt += 1
-
-#         goal.temporal_satisfied = (
-#             goal_temporal_satisfied_cnt == len(goal.temporal_vars))
-
-#     if goal.temporal_satisfied:
-#         combined_subs = {**goal.subs, **goal.new_subs}
-#         goal_temporal_vars = reify(goal.temporal_vars, combined_subs)
-
-#         if goal.obj.BaseClass is ACTION:
-#             # Unify with the KB (but for now is a simple check)
-#             # print(cycle_time, constraints_satisfied(goal), goal)
-#             if not constraints_satisfied(goal):
-
-#                 # Goal cannot be solved even after defer, discard
-#                 if goal.result is G_DEFER:
-#                     goal.update_result(G_DISCARD)
-#                     return
-
-#                 goal.clear_subs()
-#                 goal.update_result(G_CLAUSE_FAIL)
-#                 return
-
-#             if max(goal_temporal_vars) > cycle_time + 1:
-#                 goal.update_result(G_DEFER)
-#                 return
-
-#             KB.add_cycle_action(goal, combined_subs)
-
-#             goal.update_result(G_SINGLE_SOLVED)
-#             return
-#         elif goal.obj.BaseClass is EVENT:
-#             # print(goal)
-#             solve_goal(goal, cycle_time)
-#             return
-#         elif goal.obj.BaseClass is FACT:
-#             unify_fact_gen = unify_fact(goal.obj)
-#             if unify_fact_gen:
-#                 goal.set_new_subs_options(unify_fact_gen)
-#                 goal.update_subs(goal.get_new_sub_option())
-#                 goal.update_result(G_SINGLE_SOLVED)
-#                 return
-
-#             goal.clear_subs()
-#             goal.update_result(G_DISCARD)
-#             return
-#         else:
-#             raise UnhandledObjectError(goal.obj.BaseClass)
-
-#     goal.clear_subs()
-#     goal.update_result(G_UNSOLVED)
-#     return
-
-
-# def process_cycle_actions():
-#     for (action, temporal_vars) in KB.cycle_actions:
-#         KB.log_action(action, temporal_vars)
-#         causalities = KB.exists_causality(action)
-#         if causalities:
-#             for outcome in causalities.outcomes:
-#                 if outcome[0] == A_TERMINATE:
-#                     KB.remove_fluent(outcome[1])
-#                     KB.log_fluent(
-#                         outcome[1],
-#                         max(temporal_vars),
-#                         F_TERMINATE
-#                     )
-#                 elif outcome[0] == A_INITIATE:
-#                     if KB.add_fluent(outcome[1]):
-#                         KB.log_fluent(
-#                             outcome[1],
-#                             max(temporal_vars),
-#                             F_INITIATE
-#                         )
-#                 else:
-#                     raise(UnknownOutcomeError(outcome[0]))
-#     KB.clear_cycle_actions()
+'''
+Revised solver that will recursively yield solutions
+'''
+import copy
+from unification import *
+from collections import deque
+
+from pylps.constants import *
+from pylps.exceptions import *
+from pylps.utils import *
+from pylps.kb import KB
+from pylps.config import CONFIG
+
+from pylps.state import State
+
+from pylps.unifier import unify_fact
+# from pylps.solver import solve_multigoal
+from pylps.constraints import constraints_satisfied_new
+
+
+class _Solver(object):
+
+    def __init__(self):
+        self.current_time = None
+        self.cycle_actions = []
+
+    def solve_goals(self, current_time, n_solutions=1):
+        '''
+        Entry point, attempt to solve reactive rules
+        Note that reactive rules R1 .. RN are not ANDs
+        This prevents the reuse of the solver for each indivdual reactive
+        rule
+        '''
+        self.current_time = current_time
+        self.cycle_actions = []
+
+        actions_stack = []
+        states_stack = []
+        reactive_soln = {}
+
+        cur_goal_pos = 0
+        end = False
+        back = False
+
+        max_soln = 0
+        soln_cnt = 0
+        solutions = []
+
+        while not end:
+
+            # FORWARD
+            while not back:
+                if cur_goal_pos >= len(KB.goals):
+
+                    if len(self.cycle_actions) >= max_soln:
+                        # self.display_cycle_actions()
+                        solutions.append((
+                            copy.deepcopy(self.cycle_actions),
+                            copy.deepcopy(states_stack)
+                        ))
+                        max_soln = len(self.cycle_actions)
+
+                    soln_cnt += 1
+                    if soln_cnt >= n_solutions:
+                        end = True
+                        break
+
+                    back = True
+                    continue
+
+                # print('FORWARD', cur_goal_pos)
+
+                multigoal = KB.goals[cur_goal_pos]
+
+                reactive_soln[cur_goal_pos] = self.backtrack_solve(multigoal)
+
+                new_state = next(reactive_soln[cur_goal_pos])
+                self.add_cycle_actions(new_state)
+                actions_stack.append(copy.deepcopy(self.cycle_actions))
+                states_stack.append(copy.deepcopy(new_state))
+
+                cur_goal_pos += 1
+
+                # self.display_cycle_actions()
+
+            if end:
+                break
+
+            # BACK
+            while back:
+                cur_goal_pos -= 1
+                # print('BACK', cur_goal_pos)
+
+                # No more options, even for the first goal
+                if cur_goal_pos < 0:
+                    end = True
+                    break
+
+                # GO BACK TO PREV STATE
+                if actions_stack:
+
+                    actions_stack.pop()
+                    states_stack.pop()
+
+                    try:
+                        self.cycle_actions = copy.deepcopy(actions_stack[-1])
+                    except IndexError:
+                        self.cycle_actions = []
+                else:
+                    self.cycle_actions = []
+
+                try:
+                    # Do we have another solution?
+                    new_state = next(reactive_soln[cur_goal_pos])
+                    back = False
+                    self.add_cycle_actions(new_state)
+                    actions_stack.append(copy.deepcopy(self.cycle_actions))
+                    states_stack.append(copy.deepcopy(new_state))
+                    cur_goal_pos += 1
+                except StopIteration:
+                    continue
+
+        return solutions
+
+    def add_cycle_actions(self, state):
+        for action in state.actions:
+            action.args = reify_args(action.args, state.subs)
+            action.update_start_time(
+                reify_args([action.start_time], state.subs)[0])
+            action.update_end_time(
+                reify_args([action.end_time], state.subs)[0])
+
+            self.cycle_actions.append(action)
+
+    def display_cycle_actions(self):
+        for item in self.cycle_actions:
+            print(item)
+        print()
+
+    def backtrack_solve(self, start: State, pos=0):
+        '''
+        AND step of solving
+        Solve all inside here, except for the case when deferring
+        This function should return a generator
+        '''
+        states = deque()
+
+        # s_goals = deque([child.goal for child in start.children])
+        # s_subs = start.subs
+        start_state = copy.deepcopy(start)
+        states.append(start_state)
+
+        while states:
+            cur_state = states.pop()
+
+            if cur_state.result is G_DEFER:
+                yield cur_state
+                continue
+
+            # Nothing left
+            goal = cur_state.get_next_goal()
+            if not goal:
+                cur_state.set_result(G_SOLVED)
+                yield cur_state
+            else:
+                self.expand_goal(goal, cur_state, states)
+
+        '''
+        Additional empty state to cover the case where
+        the reactive rule is not chosen for solving
+        '''
+        yield State([], {})
+
+    def expand_goal(self, goal, cur_state, states):
+
+        if goal.BaseClass is ACTION:
+            self.expand_action(goal, cur_state, states)
+        elif goal.BaseClass is EVENT:
+            self.expand_event(goal, cur_state, states)
+        elif goal.BaseClass is FACT:
+            # Check if a sub is needed
+            self.expand_fact(goal, cur_state, states)
+        else:
+            raise UnimplementedOutcomeError(goal.BaseClass)
+
+    def expand_action(self, goal, cur_state, states):
+        new_state = copy.deepcopy(cur_state)
+
+        # new_state.remove_first_goal()
+
+        # Handle temporal variables (atomic action)
+        start_time = var(goal.start_time.name)
+        end_time = var(goal.end_time.name)
+
+        if start_time not in cur_state.subs:
+            # Start time has not been substituted
+            if cur_state.temporal_used:
+                print('TEMPORAL VAR USED')
+            else:
+                unify_start = unify(start_time, self.current_time)
+                unify_end = unify(end_time, self.current_time + 1)
+                new_state.update_subs(unify_start)
+                new_state.update_subs(unify_end)
+                new_state.temporal_used_true()
+        else:
+            # print('START HAS ALREADY BEEN UNIFIED')
+            # print(cur_state)
+
+            unify_end = unify(end_time, cur_state.subs[start_time] + 1)
+            new_state.update_subs(unify_end)
+            temporal_valid = new_state.subs[end_time] <= self.current_time + 1
+
+            if not temporal_valid:
+                new_state._goal_pos -= 1
+                new_state.set_result(G_DEFER)
+                states.append(new_state)
+                return
+
+        # If we execute the action, is it valid here?
+        valid = constraints_satisfied_new(
+            goal, new_state, self.cycle_actions)
+
+        # print(goal, new_state.subs, valid)
+        # print()
+
+        # Done
+        if valid:
+            new_state.add_action(goal)
+            states.append(new_state)
+
+    def expand_event(self, goal, cur_state, states):
+
+        # Need to reverse here for DFS like iteration
+        KB_clauses = list(KB.get_clauses(goal))
+        KB_clauses.reverse()
+
+        if KB_clauses:
+            if CONFIG.single_clause:
+                KB_clauses = [KB_clauses[-1]]
+
+            for clause in KB_clauses:
+                '''
+                TODO: Actually resolving the temporal requirements correctly
+                Simple replacement for now
+                '''
+                new_state = copy.deepcopy(cur_state)
+                new_state.replace_event(goal, clause.reqs)
+                states.append(new_state)
+
+    def expand_fact(self, goal, cur_state, states):
+        # Check if variables are needed
+
+        # Need to reverse here for DFS like iteration
+        subs = list(unify_fact(goal))
+        subs.reverse()
+
+        for sub in subs:
+            new_state = copy.deepcopy(cur_state)
+            # new_state.remove_first_goal()
+            new_state.update_subs(sub)
+            states.append(new_state)
+
+
+# def backtrack_solve(start, cycle_time, cycle_actions, pos=0):
+
+
+SOLVER = _Solver()
